Fixed bug with council name search box, so it now recognises councils.
Also, changed form status so it's maintained after submit button is
pressed.

Whilst editing the template to allow this to happen, I also tidied it up
a little by adding looping, and creating a Django form for the three
radio choices for how to input the council for the filter.