Suppress .is-valid classes in django_bootstrap5
It’s pretty rare that we actually want to mark form controls as “valid”
after a form is submitted. Usually, if the value of a form element is
ok, you just leave it in its default styling. (One exception, I guess,
would be when a user is choosing a password and you want to show that
they’ve met the site’s password security/complexity requirements – but
we don‘t do that either, and if we did, we’d do it live, client-side,
rather than on form submission.)

Previously we allowed django_bootstrap5 to keep adding these .is-valid
classes, but customised our Bootstrap (SCSS) settings to prevent
matching style rules from being generated in the CSS.

This solution was a kludge which meant .is-valid styles would be
unavailable to us if we ever wanted to use them (manually) in future.

Now, instead, we override django_bootstrap5.FieldRenderer to prevent it
from adding the .is-valid class to a valid field. This means we no
longer need the Bootstrap (SCSS) overrides, and we can use .is-valid
classes (manually) in future, if we want.

Part of #31, and a follow-up to #32 and #39.